break MarkdownEditorWidget into MVP
